Tuplit updated code
Fixed discount tag icon quality issue
Improved Image loading issue in merchant page
Improved performance issue in my profile page

+++ Tuplit/Tuplit/Classes/Utilities/CustomControls/ChoiceButton/ChoiceButton.m
@@ -101,7 +101,7 @@
             isSelected = NO;
         }
         
-        [delegate choiceButton:self selectedButtonIndex:btn.tag-1];
+        [delegate choiceButton:self selectedButtonIndex:(int)btn.tag-1];
         
         tag = (int)btn.tag;
     }
